+++ tests/test_pipeline.py
@@ -8,14 +8,11 @@ import pytest
 import vaex
 
 from astropy.coordinates import SkyCoord
-from astropy.io import fits
-from astropy.wcs import WCS
 from mocpy import MOC
 from pathlib import Path
-from pytest_mock import mocker
+from pytest_mock import mocker  # noqa: F401
 from typing import Dict, List, Union
 
-from vasttools import RELEASED_EPOCHS
 import vasttools.pipeline as vtp
 
 
@@ -1555,8 +1552,12 @@ class TestPipeAnalysis:
         new_image_row = dummy_PipeAnalysis.images.iloc[0]
         new_image_row.name = 10
 
-        dummy_PipeAnalysis.images = dummy_PipeAnalysis.images.append(
-            new_image_row
+        dummy_PipeAnalysis.images = pd.concat(
+            [
+                dummy_PipeAnalysis.images,
+                # need to transpose the series to a dataframe to concat
+                new_image_row.to_frame().T.set_index('name')
+            ]
         )
 
         dummy_PipeAnalysis.images.loc[10, 'skyreg_id'] = 4