+++ vasttools/query.py
@@ -11,7 +11,6 @@ import os
 import datetime
 import pandas as pd
 import warnings
-import io
 import re
 import signal
 import numexpr
@@ -21,8 +20,6 @@ import dask.dataframe as dd
 import logging
 import logging.handlers
 import logging.config
-import matplotlib.pyplot as plt
-import matplotlib.axes as maxes
 
 from astropy.coordinates import Angle
 from astropy.coordinates import SkyCoord
@@ -31,21 +28,12 @@ from astropy.nddata.utils import Cutout2D
 from astropy.coordinates import SkyCoord
 from astropy.io import fits
 from astropy.wcs import WCS
-from astropy.wcs.utils import skycoord_to_pixel
 from astropy.utils.exceptions import AstropyWarning, AstropyDeprecationWarning
-from astropy.visualization import ZScaleInterval, ImageNormalize
-from astropy.visualization import PercentileInterval
-from astropy.visualization import AsymmetricPercentileInterval
-from astropy.visualization import LinearStretch
 
 from functools import partial
 
-from matplotlib.patches import Ellipse
-from matplotlib.collections import PatchCollection
 from multiprocessing import Pool, cpu_count
 from multiprocessing_logging import install_mp_handler
-
-from mpl_toolkits.axes_grid1 import make_axes_locatable
 
 from radio_beam import Beams, Beam
 
@@ -65,7 +53,7 @@ from vasttools.survey import (
 from vasttools.source import Source
 from vasttools.utils import (
     filter_selavy_components, simbad_search, match_planet_to_field,
-    check_racs_exists, epoch12_user_warning, read_selavy
+    epoch12_user_warning, read_selavy
 )
 from vasttools.moc import VASTMOCS
 from forced_phot import ForcedPhot
@@ -1601,8 +1589,10 @@ class Query:
 
             nselavy_df[["flux_peak", "flux_int"]] *= -1.0
 
-            selavy_df = selavy_df.append(
-                nselavy_df, ignore_index=True, sort=False
+            selavy_df = pd.concat(
+                [selavy_df, nselavy_df],
+                ignore_index=True,
+                sort=False
             )
 
         selavy_coords = SkyCoord(
@@ -1626,7 +1616,7 @@ class Query:
             copy['detection'] = True
             copy['#'] = d2d.arcsec
             copy.index = group.iloc[idxc].index.values
-            master = master.append(copy, sort=False)
+            master = pd.concat([master, copy], sort=False)
             # reset index and move previous index to the end to match the meta
             master_cols = master.columns.to_list()
             master = master.reset_index()[master_cols + ['index']]
@@ -1639,7 +1629,7 @@ class Query:
             copy['detection'] = True
             copy.index = group[mask].index.values
 
-            master = master.append(copy, sort=False)
+            master = pd.concat([master, copy], sort=False)
 
             missing = group_coords[~mask]
             if missing.shape[0] > 0:
@@ -1678,7 +1668,7 @@ class Query:
 
                 rms_df.index = group[~mask].index.values
 
-                master = master.append(rms_df, sort=False)
+                master = pd.concat([master, rms_df], sort=False)
         if '#' not in master.columns:
             master.insert(0, "#", '')
         return master
@@ -1988,8 +1978,8 @@ class Query:
             if self.fields_df is None:
                 self.fields_df = planet_fields
             else:
-                self.fields_df = self.fields_df.append(
-                    planet_fields
+                self.fields_df = pd.concat(
+                    [self.fields_df, planet_fields]
                 ).reset_index(drop=True)
 
         self.logger.debug(self.fields_df)
@@ -2507,8 +2497,9 @@ class FieldQuery:
                 else:
                     to_append = load_fields_file(val)
                     to_append["EPOCH"] = epochs_dict[val]
-                    self.pilot_info = self.pilot_info.append(
-                        to_append, sort=False
+                    self.pilot_info = pd.concat(
+                        [self.pilot_info, to_append],
+                        sort=False
                     )
 
         self.field_info = self.pilot_info[